feat: rewrite module export mechanism and update tests accordingly
BREAKING CHANGE:

The module now exports a single function, which initializes the BookBrainz
ORM using a provided configuration object, and returns the initialized model
constructors and bookshelf instance. This is more elegant and future-proof
then dynamically adding models to the module itself.

To migrate, change your app to initialize bookbrainz-data-js only once, and
store the returned object, either as the export of a separate module or within
an application framework.